add ViewCharger and ViewChargers pages
